BlackBox bug fix
Fixed a bug. If the user didn't enter a "Timestamp" field as the first field, the method will write one fewer field and the header will be off. The fix checks if the first field contains "time" and if it does reformats it to "Timestamp" to make it easier for parsing programs to detect. If the user completely omits a time field, it will add one to the beginning of the header.